requirements.txt
python-dotenv~=1.1.1
ollama~=0.5.3
openai~=1.102.0